requirements.txt
tensorflow==2.0.0
setuptools==41.0.0
numpy==1.18.1
tensorflow-hub==0.7.0
django==3.0.3
djangorestframework==3.10.3